refactor: move hex_to_root beside the fixture types it parses
hex_to_root parses the hex String fields carried by the YAML fixture
types (TestMeta.genesis_validators_root, HeaderCheck.beacon_root /
execution_root), so move it from mod.rs into steps.rs next to them.
Re-export it from mod.rs so the public path test_utils::hex_to_root is
unchanged. mod.rs is now just the shared TestFork type, chain_spec, and
re-exports.
